Fixed delete feature of image based on ownership of user

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -94,16 +94,18 @@ public class ImageController {
     @RequestMapping(value = "/editImage")
     public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
 
         User user = imageService.getImage(imageId).getUser();
         User sessionUser = (User) session.getAttribute("loggeduser");
+
         if(user.getId().equals(sessionUser.getId())) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("tags", tags);
             return "images/edit";
         } else {
+            List<Tag> tags = image.getTags();
+            model.addAttribute("tags", tags);
             String error = "Only the owner of the image can edit the image";
             model.addAttribute("editError", error);
             return "images/image";
@@ -148,9 +150,23 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        Image image = imageService.getImage(imageId);
+
+        List<Tag> tags = image.getTags();
+        model.addAttribute("image", image);
+        model.addAttribute("tags", tags);
+
+        User user = imageService.getImage(imageId).getUser();
+        User sessionUser = (User) session.getAttribute("loggeduser");
+        if(user.getId().equals(sessionUser.getId())) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else {
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }
     }
